+++ Console/Assets/BlueConsole/CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## [1.1.1] - 19.11.2023
+ - Added input history recalling
  - Added a check for EventSystem in ConsoleController
 
 ## [1.1.0] - 19.11.2023
  - Added input support for old "InputSystem"
+ - Added icons to console related scripts
  - Created unity package
 
 ## [1.0.0] - 17.11.2023